fix(quadraui): backend API unit hygiene — Point cursor, f32 activity hits, f32 selection spans
Issue #504 partial pass (items 1, 3, 4 of 4):

- `EditorPaintResult::cursor_position` is now `Option<Point>` (was a
  terminal-cell `Option<(u16, u16)>` leaked from ratatui's
  `Frame::set_cursor_position`). TUI converts internally.
- `ActivityBarRowHit::{y_start, y_end}` are now `f32` (were `f64`),
  matching every other native-unit span (`Point`, `Rect`) in the crate.
  Free change: no downstream (coord-tui, vimcode) call sites reference
  the type at all (only doc comments), confirmed via
  `grep -rn ActivityBarRowHit ~/src/coord-tui/src ~/src/vimcode/src`.
- `dispatch::text_selection_line_range` now returns `Vec<(u16, f32, f32)>`
  (was `Vec<(u16, u16, u16)>`); column spans are native-unit f32 instead
  of lossy `.round() as u16` truncation, matching the "+1 unit" (not
  "+1 pixel") half-open semantics the function already relied on. Also a
  free change: zero downstream call sites (only doc mentions in vimcode).

Item 2 (TabBarHits f64 fields / tab_bar_layout_to_hits legacy
converter) deliberately deferred — it's the actual current shape of
`Backend::{draw_tab_bar, tab_bar_layout}` and friends across every
backend plus three compose/ consumers, and vimcode holds a real,
non-doc-only dependency on `TabBarHits`'s field types
(`src/core/engine/mod.rs`, `src/core/engine/terminal_ops.rs`,
`src/gtk/mod.rs`). Retiring or reshaping it is a genuine breaking
change requiring CLAUDE.md's two-PR deprecate-then-remove process and
a coordinated vimcode migration — out of scope for a single pass.
Item 5 (find_replace/status_bar/multi_section_view/split_tree u16
cells) is explicitly optional per the issue and left for follow-up.

Downstream impact: no consumer hits for any of the three changes above
(EditorPaintResult, ActivityBarRowHit, text_selection_line_range) —
grepped both ~/src/coord-tui/src and ~/src/vimcode/src. TabBarHits
(item 2, deferred) is NOT touched by this commit.

+++ quadraui/src/gtk/activity_bar.rs
@@ -236,8 +236,8 @@ pub fn draw_activity_bar_with_style(
         super::painted_text::show_layout(cr, pango_layout);
 
         regions.push(ActivityBarRowHit {
-            y_start: y,
-            y_end: y + row_h,
+            y_start: y as f32,
+            y_end: (y + row_h) as f32,
             id: item.id.clone(),
             tooltip: item.tooltip.clone(),
         });
